test: RecommendController 테스트 코드 작성 기존에 진규님이 작성하신 부분이 있어 참고하여 작성하였습니다

+++ src/main/java/org/team14/webty/recommend/repository/RecommendRepository.java
@@ -15,6 +15,7 @@ import org.team14.webty.review.entity.Review;
 
 public interface RecommendRepository extends JpaRepository<Recommend, Long> {
 	boolean existsByReviewAndUserIdAndLikeType(Review review, Long userId, LikeType likeType);
+
 	Optional<Recommend> findByReviewAndUserIdAndLikeType(Review review, Long userId, LikeType likeType);
 
 	@Query("SELECT new map( " +
@@ -25,21 +26,21 @@ public interface RecommendRepository extends JpaRepository<Recommend, Long> {
 
 	// 쿼리 7개 (JOIN FETCH 사용 X)
 	// @Query("""
-    //     SELECT r FROM Review r
-    //     JOIN Recommend rec ON r.reviewId = rec.review.reviewId
-    //     WHERE rec.userId = :userId
-    //     ORDER BY rec.voteId DESC
-    // """)
+	//     SELECT r FROM Review r
+	//     JOIN Recommend rec ON r.reviewId = rec.review.reviewId
+	//     WHERE rec.userId = :userId
+	//     ORDER BY rec.voteId DESC
+	// """)
 	// Page<Review> getUserRecommendReview(@Param("userId") Long userId, Pageable pageable);
 
 	// 쿼리 5개 (JOIN FETCH 사용 O)
 	@Query("""
-    SELECT r FROM Review r
-    JOIN FETCH r.user
-    JOIN FETCH r.webtoon
-    WHERE r.reviewId IN (SELECT rec.review.reviewId FROM Recommend rec WHERE rec.userId = :userId)
-    ORDER BY (SELECT MAX(rec.voteId) FROM Recommend rec WHERE rec.review.reviewId = r.reviewId) DESC
-""")
+		    SELECT r FROM Review r
+		    JOIN FETCH r.user
+		    JOIN FETCH r.webtoon
+		    WHERE r.reviewId IN (SELECT rec.review.reviewId FROM Recommend rec WHERE rec.userId = :userId)
+		    ORDER BY (SELECT MAX(rec.voteId) FROM Recommend rec WHERE rec.review.reviewId = r.reviewId) DESC
+		""")
 	Page<Review> getUserRecommendReview(@Param("userId") Long userId, Pageable pageable);
 
 	@Query("SELECT COALESCE(COUNT(r.voteId), 0) " +
@@ -49,11 +50,12 @@ public interface RecommendRepository extends JpaRepository<Recommend, Long> {
 	List<Long> getLikeCounts(@Param("reviewIds") List<Long> reviewIds);
 
 	@Query(value = """
-    SELECT
-        COALESCE(MAX(CASE WHEN r.like_type = 'LIKE' THEN 1 ELSE 0 END), 0) AS likes,
-        COALESCE(MAX(CASE WHEN r.like_type = 'HATE' THEN 1 ELSE 0 END), 0) AS hates
-    FROM recommend r
-    WHERE r.user_id = :userId AND r.review_id = :reviewId
-    """, nativeQuery = true)
-	Map<String, Long> findRecommendStatusByUserAndReview(@Param("userId") Long userId, @Param("reviewId") Long reviewId);
+		SELECT
+		    COALESCE(MAX(CASE WHEN r.like_type = 'LIKE' THEN 1 ELSE 0 END), 0) AS likes,
+		    COALESCE(MAX(CASE WHEN r.like_type = 'HATE' THEN 1 ELSE 0 END), 0) AS hates
+		FROM recommend r
+		WHERE r.user_id = :userId AND r.review_id = :reviewId
+		""", nativeQuery = true)
+	Map<String, Integer> findRecommendStatusByUserAndReview(@Param("userId") Long userId,
+		@Param("reviewId") Long reviewId);
 }
